moved witness utils
