+++ stwo_cairo_prover/crates/prover/src/components/mod.rs
@@ -40,7 +40,6 @@ pub mod range_check_felt_252_width_27;
 pub mod range_check_vector;
 pub mod ret_opcode;
 pub mod triple_xor_32;
-pub mod utils;
 pub mod verify_bitwise_xor_12;
 pub mod verify_bitwise_xor_4;
 pub mod verify_bitwise_xor_7;
