+++ stwo_cairo_prover/crates/prover/src/witness/blake_context.rs
@@ -6,13 +6,13 @@ use crate::cairo_air::air::CairoInteractionElements;
 use crate::cairo_air::blake::air::{
     BlakeContextClaim, BlakeContextInteractionClaim, Claim, InteractionClaim,
 };
-use crate::components::utils::TreeBuilder;
 use crate::witness::components::{
     blake_g, blake_round, blake_round_sigma, memory_address_to_id, memory_id_to_big, triple_xor_32,
     verify_bitwise_xor_12, verify_bitwise_xor_4, verify_bitwise_xor_7, verify_bitwise_xor_8,
     verify_bitwise_xor_9,
 };
 use crate::witness::range_checks::RangeChecksClaimGenerator;
+use crate::witness::utils::TreeBuilder;
 pub struct BlakeContextClaimGenerator {
     pub blake_round: blake_round::ClaimGenerator,
     pub blake_g: blake_g::ClaimGenerator,
