+++ stwo_cairo_prover/crates/prover/src/witness/opcodes.rs
@@ -5,7 +5,6 @@ use super::blake_context::BlakeContextClaimGenerator;
 use super::range_checks::RangeChecksClaimGenerator;
 use crate::cairo_air::air::CairoInteractionElements;
 use crate::cairo_air::opcodes_air::{OpcodeClaim, OpcodeInteractionClaim};
-use crate::components::utils::TreeBuilder;
 use crate::witness::components::{
     add_ap_opcode, add_ap_opcode_imm, add_ap_opcode_op_1_base_fp, add_opcode, add_opcode_imm,
     add_opcode_small, add_opcode_small_imm, assert_eq_opcode, assert_eq_opcode_double_deref,
@@ -16,6 +15,7 @@ use crate::witness::components::{
     mul_opcode_small, mul_opcode_small_imm, qm_31_add_mul_opcode, ret_opcode, verify_bitwise_xor_8,
     verify_instruction,
 };
+use crate::witness::utils::TreeBuilder;
 
 pub struct OpcodesClaimGenerator {
     add: Vec<add_opcode::ClaimGenerator>,
