+++ stwo_cairo_prover/crates/prover/src/witness/range_checks.rs
@@ -3,12 +3,12 @@ use stwo_prover::core::backend::simd::SimdBackend;
 use crate::cairo_air::range_checks_air::{
     RangeChecksClaim, RangeChecksInteractionClaim, RangeChecksInteractionElements,
 };
-use crate::components::utils::TreeBuilder;
 use crate::witness::components::{
     range_check_11, range_check_12, range_check_18, range_check_19, range_check_3_3_3_3_3,
     range_check_3_6, range_check_3_6_6_3, range_check_4_3, range_check_4_4, range_check_4_4_4_4,
     range_check_5_4, range_check_6, range_check_7_2_5, range_check_8, range_check_9_9,
 };
+use crate::witness::utils::TreeBuilder;
 
 pub struct RangeChecksClaimGenerator {
     pub rc_6_trace_generator: range_check_6::ClaimGenerator,
