+++ stwo_cairo_prover/crates/prover/src/witness/utils.rs
